(tui) cursor fixes, smooth-sb downwards
Outside any uncovered invariants, this should conclude the smooth-scroll
feature for now. Turned out quite a bit more painful than expected, but
will hopefully pay off when it comes to also being able to support the
missing advanced rendering features.

FossilOrigin-Name: b1537c5c3ffc8876249f76a2c8d88e6453f933447764e876d385e089c7d64197

+++ src/shmif/tui/tui_draw.h
@@ -310,7 +310,7 @@ static struct tui_font_ctx* tui_draw_init(size_t lim)
 static void draw_ch_u32(struct tui_font_ctx* ctx,
 	struct arcan_shmif_cont* c, uint32_t cp,
 	int x, int y, shmif_pixel fg, shmif_pixel bg,
-	size_t maxx, size_t maxy)
+	int maxx, int maxy)
 {
 	struct font_entry* font = ctx->active_font;
 	struct glyph_ent* gent;
